Dispersión o Hashing
Hashing or dispersion is a technique used in computer science to efficiently store and retrieve data in a data structure called a hash table. In C++, hashing involves converting keys into indices of an array through a hash function. This process allows for fast access to data elements based on their keys.