Integrated Project
This project encompasses a range of techniques including hashing, encryption, and decryption using Huffman coding, all implemented in C++. We utilize graphs, binary and non-binary trees, B-trees, serialization, and indexing. This comprehensive approach is tailored for a project similar to the SHEIN application. We express gratitude for the invaluable learning experiences gained through these endeavors.